+++ src/bvals/fc/bvals_fc.cpp
@@ -35,6 +35,8 @@
 #include "basic_types.hpp"
 #include "globals.hpp"
 #include "mesh/mesh.hpp"
+#include "mesh/mesh_refinement.hpp"
+#include "mesh/meshblock.hpp"
 #include "parameter_input.hpp"
 #include "utils/buffer_utils.hpp"