+++ src/outputs/parthenon_hdf5.hpp
@@ -31,7 +31,7 @@
 #include "coordinates/coordinates.hpp"
 #include "defs.hpp"
 #include "globals.hpp"
-#include "interface/container_iterator.hpp"
+#include "interface/meshblock_data_iterator.hpp"
 #include "mesh/mesh.hpp"
 #include "outputs/outputs.hpp"
 #include "parameter_input.hpp"