+++ src/utils/buffer_utils.cpp
@@ -22,6 +22,7 @@
 #include "defs.hpp"
 #include "kokkos_abstraction.hpp"
 #include "mesh/mesh.hpp"
+#include "mesh/meshblock.hpp"
 #include "parthenon_arrays.hpp"
 
 namespace parthenon {